Setting mood update

+++ app/src/main/java/pdm/moodmanager/http/user/model/UserModel.kt
@@ -14,7 +14,7 @@ class UserModel {
     var password: String = ""
 
     @SerializedName("geralScore")
-    var geralScore: Int = 0
+    var geralScore: Long = 0
 
     @SerializedName("id")
     var id: Int = 0
